Remove finishedStep from interface

+++ Components/TimestepExp/Timestep/Exponential/ArnoldiKrylov.hpp
@@ -66,7 +66,7 @@ private:
    /**
     * @brief Convergence tolerance
     */
-   const int mcTol;
+   const double mcTol;
 
    /*
     * @brief Functor for action of A matrix
@@ -115,6 +115,7 @@ int ArnoldiKrylov<TAfunc>::compute(Matrix& matV, Matrix& matH, const int jIn,
       double normV = matV.col(j+1).squaredNorm();
       // MPI VERSION HERE
       normV = std::sqrt(normV);
+
       // Stop if subspace converged sufficiently
       if(normV < this->mcTol)
       {
